Move update_weights_from_disk onto WeightUpdater
